refactor: add global exception middleware and clean try-catch from repositories

+++ Service/Interfaces/IProposalService.cs
@@ -4,7 +4,7 @@ namespace Service.Interfaces
 {
     public interface IProposalService
     {
-        Task<ProposalDto> SendProposal(int freelancerId, int jobId, int price, int hours, string message);
+        Task<ProposalDto> SendProposal(int freelancerId, int jobId, decimal price, int hours, string message);
         Task<ProposalDto> ApproveProposal(int proposalId);
         Task RejectProposal(int proposalId);
         Task<List<ProposalDto>> GetProposalsByJob(int jobId);
